feat: move python api into a separate file
